--- a/build-tools/xaprepare/xaprepare/Application/EssentialTools.RequireToolsCheck.cs
+++ b/build-tools/xaprepare/xaprepare/Application/EssentialTools.RequireToolsCheck.cs
@@ -1,0 +1,12 @@
+using System;
+
+namespace Xamarin.Android.Prepare
+{
+	partial class EssentialTools
+	{
+		bool AreToolsRequired (Context context)
+		{
+			return !context.CheckCondition (KnownConditions.EnsureEssential) && context.CheckCondition (KnownConditions.AllowProgramInstallation);
+		}
+	}
+}