--- a/build-tools/xaprepare/xaprepare/Application/EssentialTools.SevenZip.cs
+++ b/build-tools/xaprepare/xaprepare/Application/EssentialTools.SevenZip.cs
@@ -1,0 +1,21 @@
+using System;
+
+namespace Xamarin.Android.Prepare
+{
+	partial class EssentialTools
+	{
+		string sevenZipPath = String.Empty;
+
+		public string SevenZipPath {
+			get => GetRequiredValue (sevenZipPath, "7zip");
+			set => sevenZipPath = value;
+		}
+
+		partial void InitSevenZip (Context context, bool require)
+		{
+			Log.StatusLine ($"  {context.Characters.Bullet} 7za", ConsoleColor.White);
+			SevenZipPath = context.OS.Which ("7za", required: require);
+			ReportToolPath (sevenZipPath);
+		}
+	}
+}