--- a/tools/xat/Application/EssentialTools.RequireToolsCheck.cs
+++ b/tools/xat/Application/EssentialTools.RequireToolsCheck.cs
@@ -1,0 +1,10 @@
+namespace Xamarin.Android.Prepare
+{
+	partial class EssentialTools
+	{
+		bool AreToolsRequired (Context context)
+		{
+			return true;
+		}
+	}
+}