--- a/tools/xat/TestAPI/CommandFailureMode.cs
+++ b/tools/xat/TestAPI/CommandFailureMode.cs
@@ -1,0 +1,9 @@
+namespace Xamarin.Android.Tests
+{
+	enum CommandFailureMode
+	{
+		Error,
+		WarnAndContinue,
+		Quiet
+	}
+}