… build-tools/xaprepare/xaprepare/Application/ProcessRunner.cs
… build-tools/xaprepare/xaprepare/Application/Properties.Defaults.cs.in
… build-tools/xaprepare/xaprepare/Application/Utilities.Shared.cs
… build-tools/xaprepare/xaprepare/Application/Utilities.cs
… build-tools/xaprepare/xaprepare/ToolRunners/GitRunner.cs
… build-tools/xaprepare/xaprepare/ToolRunners/ToolRunner.cs
… build-tools/xaprepare/xaprepare/xaprepare.properties.projitems
… build-tools/xaprepare/xaprepare/xaprepare.targets
… tools/xat/Application/Adb.cs
… tools/xat/Application/AndroidDevice.cs
… tools/xat/Application/AppObject.cs
… tools/xat/Application/CheckAdbTarget.cs
… tools/xat/Application/Configurables.cs
… tools/xat/Application/Context.Properties.cs.in
… tools/xat/Application/Context.cs
… tools/xat/Application/CreateAndroidEmulator.cs
… tools/xat/Application/KnownProperties.cs
… tools/xat/Application/MSBuildTimingProcessor.cs
… tools/xat/Application/MSBuildTimingResult.cs
… tools/xat/Application/OS.Linux.cs
… tools/xat/Application/OS.MacOS.cs
… tools/xat/Application/OS.Unix.cs
… tools/xat/Application/OS.Windows.cs
… tools/xat/Application/OS.cs
… tools/xat/Application/ProcessLogcatTiming.cs
… tools/xat/Application/ProcessPlotInput.cs
… tools/xat/Application/RunInstrumentationTests.cs
… tools/xat/Application/RunUITests.cs
… tools/xat/Application/ShellScriptWriter.cs
… tools/xat/Application/StartAndroidEmulator.cs