`XAT` (for `Xamarin Android Testing`, imaginative) is a utility to replace our current MSBuild-based test infrastructure with a single command
that allows one to list all the tests, run all or any of them, easily select single (or more) tests from the test suites, no matter whether they're APK
or host unit tests. The intention is to make it easier for anyone not dabbling with our test suite day in day out to add, run, modify tests. The ideal
here is that one would only need to edit `tools/xat/TestCollection.cs` to do what they want.

Why is it needed? The current MSBuild architecture scatters the test running code, data describing tests across at 5 files in 4 different directories.
The `.targets` and `.projitems` files additionally reference each other causing confusion every time one (or maybe it's just me) wants to run, add or
modify any code to run and list test suites and individual tests. Additionally, wrapping everything in MSBuild causes the test run output to be 
voluminous and increases signal-to-noise ratio, making reading the output harder.

What needs reviewing the most:

 * tools/xat/TestsAPI/*
 * tools/xat/TestCollection.cs

To run the tool go to `tools/xat` and run the `run.sh` script on Unix
or, on Windows, run:

    msbuild /r
		..\..\bin\Debug\xat.exe

To see command line usage, pass the `--help` option, each sub-command
also excepts one.

What remains to be done:

 * Removal of the obsoleted MSBuild `targets` and `projitems` files (not doing it now to make review easier)
 * Implementation of test groups (to be able to say `-s NetworkTests` to run all the network tests from both APK and host test suites, for instance
 * Changes to `build-scripts/automation/` to use the new tool
 * General code polish
 * Support for .NET 6 tests