@grendello 2020-12-04
> Main thing I was wondering is if there should be `.yaml` changes here that use the new `xat` tool?
> 
> Would it simplify things like:
> 
>     * https://github.com/xamarin/xamarin-android/blob/master/build-tools/automation/yaml-templates/apk-instrumentation.yaml
> 
>     * https://github.com/xamarin/xamarin-android/blob/master/build-tools/automation/yaml-templates/run-msbuild-device-tests.yaml

It's on the TODO list (see the bottom of the PR message). I didn't want to muddle the PR with these changes now - the first thing is to get it working right, integration will come later. For the moment, it's possible to test it from command line.