@jonpryor 2021-01-16
Commit message mistake?

> To run the tool go to `tools/xat` and run the `run.sh` script on Unix

That doesn't work:

```bash
% ./tools/xat/run.sh 
Microsoft (R) Build Engine version 16.6.0 for Mono
Copyright (C) Microsoft Corporation. All rights reserved.

MSBUILD : error MSB1011: Specify which project or solution file to use because this folder contains more than one project or solution file.
```

This is presumably due to [line 2 in `run.sh`](https://github.com/xamarin/xamarin-android/pull/5369/files#diff-f8eb0c911d0f677cb8728bc3149e5bc3a8be29be6a8e43e00a045726de190ac7R2):

```bash
msbuild /r
```

If I edit `run.sh` to remove the `msbuild /r`, it doesn't work correctly if I run from "topdir":

```
% ./tools/xat/run.sh --help
Cannot open assembly '../../bin/Debug/bin/xat.exe': No such file or directory.
```

`run.sh` should presumably be:

```bash
#!/bin/bash -e

HERE="$(cd "$(dirname "$0")" && pwd)"

exec mono --debug "$HERE/../../bin/Debug/bin/xat.exe" "$@"
```

which *does* execute for me.