@jonpryor 2021-01-16
Random musings as play around with `tools/xat/run.sh`…

I would suggest attempting to make the `tools/xat/run.sh --help` output less anemic/more useful:

```
% tools/xat/run.sh --help

Locating essential tool binaries
  • git
     Found: /usr/local/bin/git
  • homebrew
     Found: /usr/local/bin/brew
  • pkgutil
     Found: /usr/sbin/pkgutil
usage: xat COMMAND [OPTIONS]

Xamarin.Android v11.2.99 test shell

Global options:
  -v, --verbosity=LEVEL      Set console log verbosity to LEVEL. Level name may
                               be abbreviated to the smallest unique part (one
                               of: silent, quiet, normal, verbose, diagnostic).
                               Default: normal
  -c, --configuration=CONFIGURATION
                             Set build CONFIGURATION instead of the default '
                               Debug'
  -D, --dull-mode            Run xat in "dull mode" - no colors, no emoji
  -E, --no-emoji             Run xat without using emoji, but still allow color

Available commands:
        list                 List known tests/categories/etc
        run                  Invoke selected tests/categories
```

I want to read top-down, assuming that information at the top is "more important".  (See also: every `man` page?)

Yet here we have a `Locating essential tool binaries` *section*, which *isn't* useful.  Then we get to a brief description, but nothing really useful.

(Actually, *every command* prints out this `Locating essential tool binaries` section, which is *useless noise*.  Unless something goes wrong, why do I want to see this?  Why should `xat list` show me the path to `git` or `brew`?)

I would suggest output similar to:

```
% run.sh --help
usage: xat COMMAND [OPTIONS]

Runs xamarin-android unit tests and applications.

See tools/xat/README.md for more information.

Global options:
  -v, --verbosity=LEVEL      Set console log verbosity to LEVEL. Level name may
                               be abbreviated to the smallest unique part (one
                               of: silent, quiet, normal, verbose, diagnostic).
                               Default: normal
  -c, --configuration=CONFIGURATION
                             Set build CONFIGURATION instead of the default '
                               Debug'
  -D, --dull-mode            Run xat in "dull mode" - no colors, no emoji
  -E, --no-emoji             Run xat without using emoji, but still allow color

Available commands:
        list                 List known tests/categories/etc
        run                  Invoke selected tests/categories

Examples:

  …

```

A month+ later, I'm still not able to remember, understand, or conceptualize the difference between a "group" and a "suite".  Perhaps different nouns should be used?

I can't quickly see what `xat list -m` *does*, other than emit more blank lines.

I wonder if a concept of "tags" would be more useful?  E.g. `AOT_Mono_Android` could have tags `{ aot Mono.Android.dll }`, `AOT_Xamarin_Forms_Performance_Integration` could have tags of `{ aot forms timing }`, etc.?

Is there/should there be a way to *search*/*filter* `xat list` output? `xat list -g`/`xat list -s` does a form of filtering, but -- as per above -- I don't *understand* this difference.

***EXAMPLES***.  The "Documentation" -- both in the commit message, and `xat --help`, and `xat run --help`, *everything* needs examples.

I should have read `tools/xat/README.md` first.  Doh.  It still lacks examples.

"test" is possibly an overused noun.  Consider the `xat run --help` output:

```
% tools/xat/run.sh run --help
…
usage: xat run [OPTIONS]
By default the command runs all the tests.
```

OK, it runs tests.

```
  -g, --group=GROUPS         Run the specified GROUPS. Takes a comma-separated
                               list of test group names or can be repeated on
                               command line
  -s, --suite=SUITES         Run the specified SUITES. Takes a comma-separated
                               list of test suite names or can be repeated on
                               command line
```

what's the relation between these, again?

```
The arguments below use suite-specific format for each entry. 
```

Oh boy.

```
… They are ignored if a test suite runs as part of predefined group:
```

I have no idea what this means. What makes a "group" a "predefined" group? Why would they be ignored?  Why should they be ignored?  If I do `xat run -g APK_all_tests`, shouldn't `-n`/`-d`/`-a` be applicable?  This header suggests ¯\_(ツ)_/¯.

```
  -t, --test=TESTS           Run the specified TESTS. Takes a comma-separated
                               list of test names
```

So run the specified tests.  What happens with `-t` is used in combination with `-g=` or `-s=`? Is it additive? Set intersection?  Other?

```
  -i, --include-categories=CATEGORIES
                             Run tests in the specified CATEGORIES. Takes a
                               comma-separated list of test categories
```

So tests have categories?  (This is presumably the NUnit/xUnit concept of categories?) How does this relate to groups & suites?  *Does* it?  *Should* it?

What *is* `Unit_No_Node`? It doesn't appear to *do* anything:

```
% tools/xat/run.sh run -g Unit_No_Node
… prints out tool paths …
```

Some of the suite names also contain "duplicative" strings, and not necessarily in a way that "conveys meaning":

  * `JavaInterop_Java_Interop_Tests`
  * `JavaInterop_Java_Interop_Dynamic_Tests`
  * `JavaInterop_Java_Interop_Export_Tests`
  * …

Both "Java" and "Interop" are repeated twice in each suite name, but what's it *mean*? It needs to be unique, certainly, but why not trim off the `JavaInterop_` prefix?  (Perhaps another place where "tags" could be a useful concept?)

I suggested earlier that "names" -- not sure if I meant group name or suite name, and I still don't know the difference -- could/should be *filesystem paths*, because then at least I *know* what it corresponds to.  I have no idea what `JavaInterop_Java_Interop_Tests` is.  There are *at least two separate projects* for "`Java.Interop-Tests`":

  * `external/Java.Interop/tests/Java.Interop-Tests`: Desktop unit tests for `Java.Interop.dll` (in `Java.Interop-Tests.dll`)
  * `tests/Mono.Android-Tests/Java.Interop-Tests`: Unit tests for `Java.Interop.dll` for use within the `Mono.Android-Tests`.

With some thought, I *think* the second one isn't "applicable", because it'll be "included" in `Mono.Android-Tests`, but this is *thinking* that I don't necessarily want to perform. Reading the source code doesn't immediately help either!

```csharp
{ "Java.Interop-Tests.dll",                            Path.Combine ("Java.Interop-Tests",                            "Java.Interop-Tests.csproj") },
```

How many `Java.Interop-Tests/Java.Interop-Tests.csproj` files are there?

```
% find . -iname Java.Interop-Tests.csproj
./tests/Mono.Android-Tests/Java.Interop-Tests/Java.Interop-Tests.csproj
./external/Java.Interop/tests/Java.Interop-Tests/Java.Interop-Tests.csproj
```

That's right, 2.

In short, the `JavaInterop_Java_Interop_Tests` suite name does me no favors, the "Host Unit" name of `JavaInterop: Java.Interop-Tests` doesn't help me, and the *source code* doesn't help me.  The only thing that *might* help, maybe, *if I got it right*, is mental effort to *exclude* one of the tests projects I know of, on the assumption that it's inclusion "doesn't make sense."

I'm thinking too hard already. :-(

What *would* be a useful "grouping" to me would be "all AOT tests".  No such grouping exists.

```
% ./tools/xat/run.sh run -s Mono_Android
…
```

What's the output look like when there's a *failure*?  Examples welcome!

As-is, output is, again, *verbose* (yay?), but not all is necessarily *useful*?

```
% ./tools/xat/run.sh run -s Mono_Android

Locating essential tool binaries
  • git
     Found: /usr/local/bin/git
  • homebrew
     Found: /usr/local/bin/brew
  • pkgutil
     Found: /usr/sbin/pkgutil
Main log file: …/xamarin-android/bin/TestDebug/tests-20210115T210001.log
Tool paths:
  • ADB: /Users/jon/android-toolchain/sdk/platform-tools/adb
  • AVD Manager: /Users/jon/android-toolchain/sdk/cmdline-tools/1.0/bin/avdmanager
  • BundleTool: …/xamarin-android/bin/Debug/lib/xamarin.android/xbuild/Xamarin/Android/bundletool.jar
  • Emulator: /Users/jon/android-toolchain/sdk/emulator/emulator
  • Java: /Users/jon/android-toolchain/jdk-11/bin/java
  • NUnit: …/xamarin-android/packages/nunit.consolerunner/3.11.1/tools/nunit3-console.exe

Running suite: Mono.Android
Running global initialization steps for suite kind 'APK', family 'APK'
 | 29
 | package:/system/priv-app/Shell/Shell.apk
 | List of devices attached
 | 98211FFBA006ND	device
 | 
Detected ADB target: 98211FFBA006ND
Building suite 'Mono.Android'
xabuild "/p:Configuration=Debug" /v:normal "/bl:../../bin/BuildDebug/msbuild-20210115T210001-build_test_Mono_Android.binlog" "/r" "/t:SignAndroidPackage" "Mono.Android-Tests.csproj" 
[MSBuild] running
[MSBuild] log file: bin/TestDebug/tests-20210115T210001.msbuild.build_test_Mono_Android.log

 | Success
Deploying package: Mono.Android_Tests-Signed.apk
 | Performing Streamed Install
 | Success
Running instrumentation tests
Starting test run for:
  Package: Mono.Android_Tests
  Instrumentation: Mono.Android_Tests/xamarin.android.runtimetests.NUnitInstrumentation
 | INSTRUMENTATION_RESULT: failed=0
 | INSTRUMENTATION_RESULT: filtered=0
 | INSTRUMENTATION_RESULT: inconclusive=0
 | INSTRUMENTATION_RESULT: nunit2-results-path=/storage/emulated/0/Android/data/Mono.Android_Tests/files/Documents/TestResults.NUnit.xml
 | INSTRUMENTATION_RESULT: passed=871
 | INSTRUMENTATION_RESULT: run=873
 | INSTRUMENTATION_RESULT: skipped=2
 | INSTRUMENTATION_RESULT: total=0
 | INSTRUMENTATION_CODE: -1
Logcat output path: …/xamarin-android/bin/TestDebug/logcat-Debug-Mono.Android_Tests.txt
Test results output path: …/xamarin-android/bin/TestDebug/TestResult-Mono.Android_Tests.xml
 | /storage/emulated/0/Android/data/Mono.Android_Tests/files/Documents/TestResults.NUnit.xml: 1 file pulled, 0 skipped. 27.7 MB/s (223350 bytes in 0.008s)
Processing timing results
   Results path: …/xamarin-android/TestResult-Mono.Android_Tests-times.csv
   Definitions path: …/xamarin-android/build-tools/scripts/TimingDefinitions.txt
Renaming test cases for 'Mono.Android': instrumentation 'xamarin.android.runtimetests.NUnitInstrumentation'
File fixed up: …/xamarin-android/TestResult-Mono.Android_Tests-Debug.xml
Running global shutdown steps for suite kind 'APK, family 'APK'

Suite execution times:
  • Mono_Android (APK): 00:00:27.1318713
Test suite succeeded
Main log file: …/xamarin-android/bin/TestDebug/tests-20210115T210001.log
```

`Locating essential tool binaries` isn't helpful. `Tool paths` isn't helpful, and tool invocations will be in the generated log files, so if they are needed, they can be obtained.


```
Running global initialization steps for suite kind 'APK', family 'APK'
 | 29
 | package:/system/priv-app/Shell/Shell.apk
 | List of devices attached
 | 98211FFBA006ND	device
```

I have no idea what this means. Is this *needed* for console output, and not log file output?

```
Building suite 'Mono.Android'
xabuild "/p:Configuration=Debug" /v:normal "/bl:../../bin/BuildDebug/msbuild-20210115T210001-build_test_Mono_Android.binlog" "/r" "/t:SignAndroidPackage" "Mono.Android-Tests.csproj" 
[MSBuild] running
[MSBuild] log file: bin/TestDebug/tests-20210115T210001.msbuild.build_test_Mono_Android.log

 | Success
```

Yay, success!  All useful!

```
Starting test run for:
  Package: Mono.Android_Tests
  Instrumentation: Mono.Android_Tests/xamarin.android.runtimetests.NUnitInstrumentation
 | INSTRUMENTATION_RESULT: failed=0
 | INSTRUMENTATION_RESULT: filtered=0
 | INSTRUMENTATION_RESULT: inconclusive=0
 | INSTRUMENTATION_RESULT: nunit2-results-path=/storage/emulated/0/Android/data/Mono.Android_Tests/files/Documents/TestResults.NUnit.xml
 | INSTRUMENTATION_RESULT: passed=871
 | INSTRUMENTATION_RESULT: run=873
 | INSTRUMENTATION_RESULT: skipped=2
 | INSTRUMENTATION_RESULT: total=0
 | INSTRUMENTATION_CODE: -1
Logcat output path: /Volumes/Xamarin-Work/xamarin-android/bin/TestDebug/logcat-Debug-Mono.Android_Tests.txt
Test results output path: /Volumes/Xamarin-Work/xamarin-android/bin/TestDebug/TestResult-Mono.Android_Tests.xml
 | /storage/emulated/0/Android/data/Mono.Android_Tests/files/Documents/TestResults.NUnit.xml: 1 file pulled, 0 skipped. 27.7 MB/s (223350 bytes in 0.008s)
```

Notably absent from this section: "Success", or any "clear" indicator that it was successful, *unless* you *know* that `INSTRUMENTATION_CODE: -1` means "👍". An explicit "Success" message might be nice here?


```
Processing timing results
   Results path: /Volumes/Xamarin-Work/xamarin-android/TestResult-Mono.Android_Tests-times.csv
   Definitions path: /Volumes/Xamarin-Work/xamarin-android/build-tools/scripts/TimingDefinitions.txt
Renaming test cases for 'Mono.Android': instrumentation 'xamarin.android.runtimetests.NUnitInstrumentation'
File fixed up: /Volumes/Xamarin-Work/xamarin-android/TestResult-Mono.Android_Tests-Debug.xml
Running global shutdown steps for suite kind 'APK, family 'APK'
```

Not sure if it's useful for console output.  `Main log file` contains all the same information and more; perhaps "Processing timing results" doesn't *need* to be in the console output?